pr feedback structure
