+++ client-react/src/pages/app/functions/function/function-log/FunctionLogFileStreamData.ts
@@ -1,6 +1,6 @@
 import { LogEntry, LogLevel } from './FunctionLog.types';
 import { CommonConstants } from '../../../../../utils/CommonConstants';
-import { LogRegex, LogLevel as FBLogLevel, maxLogEntries } from '../../../log-stream/web-app/LogStream.types';
+import { LogRegex, LogLevel as FBLogLevel, maxLogEntries } from '../../../log-stream/LogStream.types';
 
 export function processLogs(logStream: string, oldLogs: LogEntry[]): LogEntry[] {
   let updatedLogs = oldLogs;
