+++ client-react/src/pages/app/functions/log-stream/LogStream.styles.ts
