+++ client-react/src/pages/app/log-stream/LogStreamData.ts
@@ -1,7 +1,7 @@
 import { LogEntry, newLine, LogLevel, LogRegex, maxLogEntries, LogsEnabled, LogType } from './LogStream.types';
-import { TextUtilitiesService } from '../../../../utils/textUtilities';
-import { Site } from '../../../../models/site/site';
-import { SiteLogsConfig } from '../../../../models/site/logs-config';
+import { TextUtilitiesService } from '../../../utils/textUtilities';
+import { Site } from '../../../models/site/site';
+import { SiteLogsConfig } from '../../../models/site/logs-config';
 
 export function processLogs(logStream: string, oldLogs: LogEntry[]): LogEntry[] {
   let updatedLogs = oldLogs;
