+++ client-react/src/pages/app/log-stream/LogStreamLogContainer.tsx
@@ -3,12 +3,12 @@ import { useTranslation } from 'react-i18next';
 import { style } from 'typestyle';
 import { LogEntry, LogType, LogsEnabled } from './LogStream.types';
 import { ChoiceGroup, IChoiceGroupOption } from 'office-ui-fabric-react';
-import { ScenarioService } from '../../../../utils/scenario-checker/scenario.service';
-import { ScenarioIds } from '../../../../utils/scenario-checker/scenario-ids';
+import { ScenarioService } from '../../../utils/scenario-checker/scenario.service';
+import { ScenarioIds } from '../../../utils/scenario-checker/scenario-ids';
 import { getLogTextColor } from './LogStreamData';
-import { ChoiceGroupStyles } from '../../../../theme/CustomOfficeFabric/AzurePortal/ChoiceGroup.styles';
-import { ArmObj } from '../../../../models/arm-obj';
-import { Site } from '../../../../models/site/site';
+import { ChoiceGroupStyles } from '../../../theme/CustomOfficeFabric/AzurePortal/ChoiceGroup.styles';
+import { ArmObj } from '../../../models/arm-obj';
+import { Site } from '../../../models/site/site';
 
 interface LogStreamLogContainerProps {
   clearLogs: boolean;
